jour 2

+++ database/migrations/2026_05_26_102127_create_order_items_table.php
@@ -5,20 +5,18 @@ use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration {
-    public function up(): void
-    {
+    public function up(): void {
         Schema::create('order_items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('order_id')->constrained()->onDelete('cascade');
             $table->foreignId('dish_id')->constrained()->onDelete('cascade');
             $table->integer('quantity');
-            $table->double('price'); 
+            $table->decimal('price', 10, 2); // Prix unitaire au moment de la commande
             $table->timestamps();
         });
     }
 
-    public function down(): void
-    {
+    public function down(): void {
         Schema::dropIfExists('order_items');
     }
 };
